+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.1.17 (under development)
 
-- Added `Entry::validateSlug` to check against real URL paths
+- Added `Entry::validateSlug()` to check against real URL paths
+- Added `SetupController::ensureFolder()` to create missing folder on setup
 
 ## 2.1.16 (Jan 26, 2024)